CVlm {DAAG}	R Documentation

Cross-Validation for Linear Regression

Description

This function gives internal and cross-validation measures of predictive accuracy for ordinary linear regression. The data are randomly assigned to a number of `folds'. Each fold is removed, in turn, while the remaining data is used to re-fit the regression model and to predict at the deleted observations.

Usage

CVlm(df = houseprices, form.lm = formula(sale.price ~ area), m=3, dots = 
FALSE, seed=29, plotit=TRUE, printit=TRUE)

Arguments

df 	 a data frame

form.lm 	 a formula object

m 	 the number of folds

dots 	 uses pch=16 for the plotting character

seed 	 random number generator seed

plotit 	 if TRUE, a plot is constructed on the active device

printit 	 if TRUE, output is printed to the screen

Value

For each fold, a table listing

the residuals
ms = the overall mean square of prediction error

Author(s)

J.H. Maindonald

See Also

lm

Examples

CVlm()

[Package DAAG version 0.81 Index]