ais {DAAG}	R Documentation

Australian athletes data set

Description

These data were collected in a study of how data on various characteristics of the bloood varied with sport body size and sex of the athlete.

Usage

data(ais)

Format

A data frame with 202 observations on the following 13 variables.

rcc

red blood cell count, in 10^{12} l^{-1}

wcc

while blood cell count, in 10^{12} per liter

hc

hematocrit, percent

hg

hemaglobin concentration, in g per decaliter

ferr

plasma ferritins, ng dl^{-1}

bmi

Body mass index, kg cm^{-2} 10^2

ssf

sum of skin folds

pcBfat

percent Body fat

lbm

lean body mass, kg

ht

height, cm

wt

weight, kg

sex

a factor with levels f m

sport

a factor with levels B_Ball Field Gym Netball Row Swim T_400m T_Sprnt Tennis W_Polo

Details

Do blood hemoglobin concentrations of athletes in endurance-related events differ from those in power-related events?

Source

These data were the basis for the analyses that are reported in Telford and Cunningham (1991).

References

Telford, R.D. and Cunningham, R.B. 1991. Sex, sport and body-size dependency of hematology in highly trained athletes. Medicine and Science in Sports and Exercise 23: 788-794.

[Package DAAG version 0.81 Index]