bestsetNoise {DAAG}	R Documentation

Best Subset Selection Applied to Noise

Description

Best subset selection applied to completely random noise. This function demonstrates how variable selection techniques in regression can often err in suggesting that more variables be included in a regression model than necessary.

Usage

bestsetNoise(m=100, n=40, method="exhaustive", nvmax=3)

Arguments

m 	 the number of observations to be simulated.

n 	 the number of predictor variables in the simulated model.

method 	 Use exhaustive search, or backward selection, or forward selection, or sequential replacement.

nvmax 	 maximum number of explanatory variables in model.

Details

A set of n predictor variables are simulated as independent standard normal variates, in addition to a response variable which is also independent of the predictors. The best model with nvmax variables is selected using the regsubsets() function from the leaps package. (The leaps package must be installed for this function to work.)

Value

bestsetNoise returns the lm model object for the "best" model.

Author(s)

J.H. Maindonald

See Also

lm

Examples

leaps.out <- try(require(leaps, quietly=TRUE))
leaps.out.log <- is.logical(leaps.out)
if ((leaps.out.log==TRUE)&(leaps.out==TRUE))
bestsetNoise(20,6) # `best' 3-variable regression for 20 simulated observations 
                   # on 7 unrelated variables (including the response)

[Package DAAG version 0.95 Index]