biomass {DAAG}	R Documentation

Biomass Data

Description

The biomass data frame has 135 rows and 8 columns. The rainforest data frame is a subset of this one.

Usage

biomass

Format

This data frame contains the following columns:

dbh

a numeric vector

wood

a numeric vector

bark

a numeric vector

fac26

a factor with 3 levels

root

a numeric vector

rootsk

a numeric vector

branch

a numeric vector

species

a factor with levels Acacia mabellae, C. fraseri, Acmena smithii, B. myrtifolia

Source

J. Ash, Australian National University

References

Ash, J. and Helman, C. (1990) Floristics and vegetation biomass of a forest catchment, Kioloa, south coastal N.S.W. Cunninghamia, 2: 167-182.

[Package DAAG version 0.81 Index]