capstring {DAAG}	R Documentation

Converts initial character of a string to upper case

Description

This function is useful for use before plotting, if one wants capitalized axis labels or factor levels.

Usage

capstring(names)

Arguments

names 	 a character vector

Value

a character vector with upper case initial values

Author(s)

W.J. Braun

Examples

capstring(names(tinting)[c(3,4)])

library(lattice)
levels(tinting$agegp) <- capstring(levels(tinting$agegp))
xyplot(csoa ~ it | sex * agegp, data=tinting) 

[Package DAAG version 0.81 Index]