component.residual {DAAG}	R Documentation

Component + Residual Plot

Description

Component + Residual plot for a term in a lm model.

Usage

component.residual(lm.obj = mice12.lm, which = 1, xlab = "Component",
    ylab = "C+R")

Arguments

lm.obj 	 A lm object

which 	 numeric code for the term in the lm formula to be plotted

xlab 	 label for the x-axis

ylab 	 label for the y-axis

Value

A scatterplot with a smooth curve overlaid.

Author(s)

J.H. Maindonald

See Also

lm

Examples

mice12.lm <- lm(brainwt ~ bodywt + lsize, data=litters)
oldpar <- par(mfrow = c(1,2))
component.residual(mice12.lm, 1, xlab = "Body weight", ylab= "t(Body weight) + e")
component.residual(mice12.lm, 2, xlab = "Litter size", ylab= "t(Litter size) + e")
par(oldpar)

[Package DAAG version 0.81 Index]