datafile {DAAG}	R Documentation

Create an ASCII data file

Description

Invoking this function creates one of four data files used in Chapters 1 and 14 of DAAGUR.

Usage

datafile(file="fuel")

Arguments

file 	 character; "fuel", for fuel.txt; "fuel.csv", for fuel.csv; "oneBadRow", for oneBadRow.txt; "scan-demo", for scan-demo.txt.

Value

One of four ASCII files is produced, and output to the current working directory.

Author(s)

J.H. Maindonald

[Package DAAG version 0.81 Index]