fossilfuel {DAAG}	R Documentation

Fossil Fuel Data

Description

Usage

fossilfuel

Format

This data frame contains the following columns:

year

a numeric vector giving the year the measurement was taken.

carbon

a numeric vector giving the total worldwide carbon emissions from fossil fuel use, in millions of tonnes.

Source

Marland et al (2003)

Examples

plot(fossilfuel)

[Package DAAG version 0.81 Index]