frostedflakes {DAAG}	R Documentation

Frosted Flakes data

Description

The frosted flakes data frame has 101 rows and 2 columns giving the sugar concentration (in percent) for 25 g samples of a cereal as measured by 2 methods – high performance liquid chromatography (a slow accurate lab method) and a quick method using the infra-analyzer 400.

Usage

elastic1

Format

This data frame contains the following columns:

Lab

careful laboratory analysis measurements using high performance liquid chromatography

IA400

measurements based on the infra-analyzer 400

Source

W. J. Braun

[Package DAAG version 0.88 Index]