hardcopy {DAAG}	R Documentation

Graphical Output for Hardcopy

Description

This function streamlines graphical output to the screen, pdf or ps files.

Usage

hardcopy(width=3.75, height=3.75, color=F, trellis=F,
             device=c("","pdf","ps"), path="", pointsize=c(8,4), horiz=F)

Arguments

width

height

color 	 TRUE if plot is not black on white only

trellis 	 TRUE if plot uses trellis graphics

device 	 screen "", pdf or ps

path 	 external file name

pointsize

horiz 	 FALSE for landscape mode

Value

Graphical output to screen, pdf or ps file.

Author(s)

J.H. Maindonald

See Also

postscript

[Package DAAG version 0.81 Index]