headInjury {DAAG}	R Documentation

Minor Head Injury (Simulated) Data

Description

The headInjury data frame has 3121 rows and 11 columns. The data were simulated according to a simple logistic regression model to match roughly the clinical characteristics of a sample of individuals who suffered minor head injuries.

Usage

headInjury

Format

This data frame contains the following columns:

age.65

age factor (0 = under 65, 1 = over 65).

amnesia.before

amnesia before impact (less than 30 minutes = 0, more than 30 minutes =1).

basal.skull.fracture

(0 = no fracture, 1 = fracture).

GCS.decrease

Glasgow Coma Scale decrease (0 = no deterioration, 1 = deterioration).

GCS.13

initial Glasgow Coma Scale (0 = not `13', 1 = `13').

GCS.15.2hours

Glasgow Coma Scale after 2 hours (0 = not `15', 1 = '15').

high.risk

assessed by clinician as high risk for neurological intervention (0 = not high risk, 1 = high risk).

loss.of.consciousness

(0 = conscious, 1 = loss of consciousness).

open.skull.fracture

(0 = no fracture, 1 = fracture)

vomiting

(0 = no vomiting, 1 = vomiting)

clinically.important.brain.injury

any acute brain finding revealed on CT (0 = not present, 1 = present).

References

Stiell, I.G., Wells, G.A., Vandemheen, K., Clement, C., Lesiuk, H., Laupacis, A., McKnight, R.D., Verbee, R., Brison, R., Cass, D., Eisenhauer, M., Greenberg, G.H., and Worthington, J. (2001) The Canadian CT Head Rule for Patients with Minor Head Injury, The Lancet. 357: 1391-1396.

[Package DAAG version 0.81 Index]