leafshape {DAAG}	R Documentation

Full Leaf Shape Data Set

Description

Leaf length, width and petiole measurements taken at various sites in Australia.

Usage

leafshape

Format

This data frame contains the following columns:

bladelen

leaf length (in mm)

petiole

a numeric vector

bladewid

leaf width (in mm)

latitude

latitude

logwid

natural logarithm of width

logpet

logarithm of petiole

loglen

logarithm of length

arch

leaf architecture (0 = plagiotropic, 1 = orthotropic

location

a factor with levels
Sabah, Panama, Costa Rica, N Queensland, S Queensland, Tasmania

Source

King, D.A. and Maindonald, J.H. 1999. Tree architecture in relation to leaf dimensions and tree stature in temperate and tropical rain forests. Journal of Ecology 87: 1012-1024.

[Package DAAG version 0.81 Index]