lmdiags {DAAG}	R Documentation

Return data required for diagnostic plots

Description

This extracts the code that provides the major part of the statistical information used by plot.lm, leaving out the code used to provide the graphs

Usage

lmdiags(x, which = c(1L:3L, 5L), cook.levels = c(0.5, 1), hii=NULL)

Arguments

x
This must be an object of class lm object, or that inherits from an object of class lm.

which
a subset of the numbers '1:6', indicating the plots for which statistical information is required

cook.levels
Levels for contours of cook.levels, by default c(0.5,1)

hii
Diagonal elements for the hat matrix. If not supplied (hii=NULL), they will be calculated from the argument x.

Details

See plot.lm for additional information.

Value

yh
fitted values

rs
standardized residuals (for glm models standardized deviance residuals)

yhn0
As yh, but omitting observations with zero weight

cook
Cook's statistics

rsp
standardized residuals (for glm models standardized Pearson residuals)

Note

This function is designed, in the first place, for use in connection with plotSimDiags, used to give simulations of diagnostic plots for lm objects.

Author(s)

John Maindonald, using code that John Maindonald, Martin Maechler and others had contributed to plot.lm

References

See references for plot.lm

See Also

plotSimDiags, plot.lm

Examples

women.lm <- lm(weight ~ height, data=women)
veclist <- lmdiags(x=women.lm)
## Returns the statistics that are required for graphs 1, 2, 3, and 5

[Package DAAG version 1.25.4 Index]