lung {DAAG}	R Documentation

Cape Fur Seal Lung Measurements

Description

The lung vector consists of weight measurements of lungs taken from 30 Cape Fur Seals that died as an unintended consequence of commercial fishing.

Usage

lung

[Package DAAG version 0.95 Index]