nsw74demo {DAAG}	R Documentation

Labour Training Evaluation Data

Description

This data frame contains 445 rows and 10 columns. These data are from an investigation of the effect of training on changes, between 1974-1975 and 1978, in the earnings of individuals who had experienced employment difficulties Data are for the male experimental control and treatment groups.

Usage

nsw74demo

Format

This data frame contains the following columns:

trt

a numeric vector identifying the study in which the subjects were enrolled (0 = PSID, 1 = NSW).

age

age (in years).

educ

years of education.

black

(0 = not black, 1 = black).

hisp

(0 = not hispanic, 1 = hispanic).

marr

(0 = not married, 1 = married).

nodeg

(0 = completed high school, 1 = dropout).

re74

real earnings in 1974.

re75

real earnings in 1975.

re78

real earnings in 1978.

Source

http://www.columbia.edu/~rd247/nswdata.html

References

Dehejia, R.H. and Wahba, S. 1999. Causal effects in non-experimental studies: re-evaluating the evaluation of training programs. Journal of the American Statistical Association 94: 1053-1062.

Lalonde, R. 1986. Evaluating the economic evaluations of training programs. American Economic Review 76: 604-620.

[Package DAAG version 0.95 Index]