onesamp {DAAG}	R Documentation

Paired Sample t-test

Description

This function performs a t-test for the mean difference for paired data, and produces a scatterplot of one column against the other column, showing whether there was any benefit to using the paired design.

Usage

onesamp(dset=corn, x="unsprayed", y="sprayed", xlab=NULL, ylab=NULL, dubious=NULL, conv=NULL, dig=2)

Arguments

dset 	 a matrix or dataframe having two columns

x 	 name of column to play the role of the `predictor'

y 	 name of column to play the role of the `response'

xlab 	 horizontal axis label

ylab 	 vertical axis label

dubious

conv

dig

Value

A scatterplot of y against x together with estimates of standard errors and standard errors of the difference (y-x).
Also produced is a confidence interval and p-value for the test.

Author(s)

J.H. Maindonald

Examples

onesamp(dset = pair65, x = "ambient", y = "heated", xlab =
        "Amount of stretch (ambient)", ylab =
        "Amount of stretch (heated)") 

[Package DAAG version 0.81 Index]