onewayPlot {DAAG}	R Documentation

Display of One Way Analysis Results

Description

A line plot of estimates for unstructured comparison of factor levels

Usage

onewayPlot(obj, trtnam = "trt", axisht = 4.5, xlim = NULL,
    xlab = NULL, lsdht = 1.5, hsdht = 0.5, textht = axisht -
        2.5, oma = rep(1, 4), angle = 80, alpha = 0.05)
oneway.plot(obj, trtnam = "trt", axisht = 4.5, xlim = NULL,
    xlab = NULL, lsdht = 1.5, hsdht = 0.5, textht = axisht -
        2.5, oma = rep(1, 4), angle = 80, alpha = 0.05)

Arguments

obj
One way analysis of variance object (from aov)

trtnam
name of factor for which line plot is required

axisht
Axis height

xlim
Range on horizontal axis

xlab
Horizontal axis label

lsdht
Height adjustment parameter for display of LSD

hsdht
Height adjustment parameter for display of Tukey's HSD

textht
Height of text

oma
Outer margin area

angle
Text angle (in degrees)

alpha
Test size

Value

Estimates, labeled with level names, are set out along a line

Author(s)

J.H. Maindonald

Examples

rice.aov <- aov(ShootDryMass ~ trt, data=rice)
onewayPlot(obj=rice.aov)

[Package DAAG version 1.25.4 Index]