overlap.density {DAAG}	R Documentation

Overlapping Density Plots - obsolete

Description

Densities for two independent samples are estimated and plotted.

Usage

overlap.density(x0, x1, ratio=c(0.05, 20), compare.numbers=TRUE, 
plotit=TRUE, gpnames=c("Control", "Treatment"), xlab="Score")

Arguments

x0 	 control group measurements

x1 	 treatment group measurements

ratio 	 the range within which the relative numbers of observations from the two groups are required to lie. [The relative numbers at any point are estimated from (density1*n1)/(density0*x0)]

compare.numbers 	 If TRUE (default), then density plots are scaled to have total area equal to the sample size; otherwise total area under each density is 1

plotit 	 If TRUE, a plot is produced

gpnames 	 Names of the two samples

xlab 	 Label for x-axis

Author(s)

J.H. Maindonald

See Also

t.test

Examples

attach(two65)
overlap.density(ambient,heated)
t.test(ambient,heated)

[Package DAAG version 0.81 Index]