overlapDensity {DAAG}	R Documentation

Overlapping Density Plots

Description

Densities for two independent samples are estimated and plotted.

Usage

overlapDensity(x0, x1, ratio = c(0.05, 20), compare.numbers = FALSE,
            plotit = TRUE, gpnames = c("Control", "Treatment"),
            cutoffs=c(lower=TRUE, upper=TRUE), bw=FALSE,
            xlab = "Score", col=1:2, lty=1:2)

Arguments

x0 	 control group measurements

x1 	 treatment group measurements

ratio 	 the range within which the relative numbers of observations from the two groups are required to lie. [The relative numbers at any point are estimated from (density1*n1)/(density0*x0)]

compare.numbers 	 If TRUE (default), then density plots are scaled to have total area equal to the sample size; otherwise total area under each density is 1

plotit 	 If TRUE, a plot is produced

gpnames 	 Names of the two samples

cutoffs 	 logical vector, indicating whether density estimates should be truncated below (lower=TRUE) or above (upper=TRUE)

bw 	 logical, indicates whether to overwrite with a gray scale plot

xlab 	 Label for x-axis

col 	 standard color parameter

lty 	 standard line type preference

Author(s)

J.H. Maindonald

See Also

t.test

Examples

attach(two65)
overlapDensity(ambient,heated)
t.test(ambient,heated)

[Package DAAG version 0.81 Index]