possumsites {DAAG}	R Documentation

Possum Sites

Description

The possumsites data frame consists of latitudes, longitudes, and altitudes for the seven sites from Southern Victoria to central Queensland where the possum observations were made.

Usage

possumsites

Format

This data frame contains the following columns:

latitude

a numeric vector

longitude

a numeric vector

altitude

in meters

Source

Lindenmayer, D. B., Viggers, K. L., Cunningham, R. B., and Donnelly, C. F. 1995. Morphological variation among columns of the mountain brushtail possum, Trichosurus caninus Ogilby (Phalangeridae: Marsupiala). Australian Journal of Zoology 43: 449-458.

Examples

require(oz)
oz(sections=c(3:5, 11:16))
attach(possumsites)
points(latitude, longitude, pch=16, col=2)
chw <- par()$cxy[1]
chh <- par()$cxy[2]
posval <- c(2,4,2,2,4,2,2)
text(latitude+(3-posval)*chw/4, longitude, row.names(possumsites), pos=posval)

[Package DAAG version 0.81 Index]