sumry.glm {DAAG}	R Documentation

Summarizing Generalized Linear Model Fits

Description

These functions are methods for class glm or sumry.glm objects.

Usage

## S3 method for class 'glm'
sumry(object, dispersion = NULL, correlation = FALSE,
        symbolic.cor = FALSE, ...)

## S3 method for class 'sumry.glm'
print(x, digits = max(3L, getOption("digits") - 3L), 
                          symbolic.cor = FALSE, 
                          signif.stars = getOption("show.signif.stars"), 
                          call=FALSE, deviance.residuals=FALSE, 
                          show.iter=10, ...) 

Arguments

object
an object of class "glm", usually, a result of a call to glm.

x
an object of class "summary.glm", usually, a result of a call to summary.glm.

dispersion
the dispersion parameter for the family used. Either a single numerical value or NULL (the default), when it is inferred from object (see ‘Details’).

correlation
logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.

digits
the number of significant digits to use when printing.

symbolic.cor
logical. If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.

signif.stars
logical. If TRUE, ‘significance stars’ are printed for each coefficient.

call
logical. If TRUE, details of the function call are printed.

deviance.residuals
logical. If TRUE, deviance residuals are printed.

show.iter
NULL or integer. If NULL, or if the number of iterations is greater than the specified integer, then the number of iterations will be printed.

...
further arguments passed to or from other methods.

Details

The function print.sumry.glm allows, relative to print.summary.glm, some greater flexibility in what is printed. By default, details of the call to glm are omitted, and details of the number of interations only in the unusual case where this number is greater than 10. See the help page for summary.glm for further details.

Value

sumry.glm returns an object of class "sumry.glm", a list with the same components as summary.glm.

See Also

glm, summary.

Examples

## For examples see example(glm)

[Package DAAG version 1.25.4 Index]