simulateLinear {DAAG}	R Documentation

Simulation of Linear Models for ANOVA vs. Regression Comparison

Description

This function simulates a number of bivariate data sets in which there are replicates at each level of the predictor. The p-values for ANOVA and for the regression slope are compared.

Usage

simulateLinear(sd=2, npoints=5, nrep=4, nsets=200, type="xy", seed=21)

Arguments

sd 	 The error standard deviation

npoints 	 Number of distinct predictor levels

nrep 	 Number of replications at each level

nsets 	 Number of simulation runs

type 	 Type of data

seed 	 Random Number generator seed

Value

The proportion of regression p-values that are less than the ANOVA p-values is printed

Author(s)

J.H. Maindonald

Examples

simulateLinear()

[Package DAAG version 0.81 Index]