stVincent {DAAG}	R Documentation

Averages by block of yields for the St. Vincent Corn data

Description

These data frames have yield averages by blocks (parcels).

Usage

stVincent

Format

A data frame with 324 observations on 8 variables.

code

a numeric vector

island

a numeric vector

id

a numeric vector

site

a factor with 8 levels.

block

a factor with levels I II III IV

plot

a numeric vector

trt

a factor consisting of 12 levels

harvwt

a numeric vector; the average yield

Source

Andrews DF; Herzberg AM, 1985. Data. A Collection of Problems from Many Fields for the Student and Research Worker. Springer-Verlag. (pp. 339-353)

[Package DAAG version 0.81 Index]