twotPermutation {DAAG}	R Documentation

Two Sample Permutation Test

Description

This function computes the p-value for the two sample t-test using a permutation test. The permutation density can also be plotted.

Usage

twotPermutation(x1=two65$ambient, x2=two65$heated, nsim=2000, plotit=TRUE)

Arguments

x1 	 Sample 1

x2 	 Sample 2

nsim 	 Number of simulations

plotit 	 If TRUE, the permutation density will be plotted

Details

Suppose we have n1 values in one group and n2 in a second, with n = n1 + n2. The permutation distribution results from taking all possible samples of n2 values from the total of n values.

Value

The p-value for the test of the hypothesis that the mean of x1 differs from x2

Author(s)

J.H. Maindonald

References

Good, P. 2000. Permutation Tests. Springer, New York.

Examples

twotPermutation()

[Package DAAG version 0.81 Index]