vince111b {DAAG}	R Documentation

Averages by block of corn yields, for treatment 111 only

Description

These data frames have averages by blocks (parcels) for the treatment 111.

Usage

vince111b

Format

A data frame with 36 observations on 8 variables.

site

a factor with levels AGSV CASV CPSV LPSV MPSV OOSV OTSV SSSV UISV

parcel

a factor with levels I II III IV

code

a numeric vector

island

a numeric vector

id

a numeric vector

plot

a numeric vector

trt

a numeric vector

harvwt

a numeric vector

Source

Andrews DF; Herzberg AM, 1985. Data. A Collection of Problems from Many Fields for the Student and Research Worker. Springer-Verlag. (pp. 339-353)

[Package DAAG version 0.81 Index]