whoops {DAAG}	R Documentation

Deaths from whooping cough, in London

Description

Deaths from whooping cough, in London from 1740 to 1881.

Usage

data(whoops)

Format

This is a multiple time series consisting of 3 series: wcough, ratio, and alldeaths.

Source

Guy, W. A. 1882. Two hundred and fifty years of small pox in London. Journal of the Royal Statistical Society 399-443.

References

Lancaster, H. O. 1990. Expectations of Life. Springer.

Examples

data(whoops)
str(whoops)
plot(whoops) 

[Package DAAG version 0.81 Index]