fossilfuel {DAAG}	R Documentation

Fossil Fuel Data

Description

Estimates of total worldwide carbon emissions from fossil fuel use.

Usage

fossilfuel

Format

This data frame contains the following columns:

year

a numeric vector giving the year the measurement was taken.

carbon

a numeric vector giving the total worldwide carbon emissions from fossil fuel use, in millions of tonnes.

Details

Data for the years 1751 through to 2014 is available from Data for the years 2014 https://cdiac.ess-dive.lbl.gov/ftp/ndp030/global.1751_2014.ems

Source

Boden TA, Marland G, Andres RJ (2017). “Global, Regional, and National Fossil-Fuel CO2 Emissions.” https://cdiac.ess-dive.lbl.gov/trends/emis/meth_reg.html.

Examples

plot(fossilfuel)

[Package DAAG version 1.25.4 Index]