pause {DAAG}	R Documentation

Pause before continuing execution

Description

If a program produces several plots, isertion of pause() between two plots suspends execution until the <Enter> key is pressed, to allow inspection of the current plot.

Usage

pause()

Author(s)

From the ‘sm’ package of Bowman and Azzalini (1997)

[Package DAAG version 1.25.4 Index]