simulateLinear {DAAG}	R Documentation

Simulation of Linear Models for ANOVA vs. Regression Comparison

Description

This function simulates a number of bivariate data sets in which there are replicates at each level of the predictor. The p-values for ANOVA and for the regression slope are compared, and a lattice graphics object returned.

Usage

simulateLinear(sd=2, npoints=5, nrep=4, nsets=200, graphtype="xy",
seed=21, ...)

Arguments

sd
The error standard deviation

npoints
Number of distinct predictor levels

nrep
Number of replications at each level

nsets
Number of simulation runs

graphtype
Type of graph; x-y plot (graphtype="xy"), overlaid density plots (graphtype="density"), or density plot for x-y difference (graphtype="density-diff")

seed
Random Number generator seed

...
Additional arguments, to be passed through to the lattice function that is called

Value

A lattice graphics object.

Author(s)

J.H. Maindonald

Examples

simulateLinear()

[Package DAAG version 1.25.4 Index]