vif {DAAG}	R Documentation

Variance Inflation Factors

Description

Variance inflation factors are computed for the standard errors of linear model coefficient estimates.

Usage

vif(obj, digits=5)

Arguments

obj
A lm object

digits
Number of digits

Value

A vector of variance inflation factors corresponding to the coefficient estimates given in the lm object.

Author(s)

J.H. Maindonald

See Also

lm

Examples

litters.lm <- lm(brainwt ~ bodywt + lsize, data = litters)
vif(litters.lm)

carprice1.lm <- lm(gpm100 ~ Type+Min.Price+Price+Max.Price+Range.Price,
    data=carprice)
vif(carprice1.lm)

carprice.lm <- lm(gpm100 ~ Type + Price, data = carprice)
vif(carprice1.lm)

[Package DAAG version 1.25.4 Index]